% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{as_ggplot}
\alias{as_ggplot}
\alias{as_ggplot.ggplot}
\alias{as_ggplot.kggplot}
\title{Convert to a ggplot2 object}
\usage{
as_ggplot(x, ...)

\method{as_ggplot}{ggplot}(x, ...)

\method{as_ggplot}{kggplot}(x, ...)
}
\arguments{
\item{x}{A \code{kggplot} (or a \code{ggplot}, returned unchanged).}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{as_ggplot()} renders a \code{\link[=kggplot]{kggplot()}} specification into a regular
\code{ggplot} object, which you can then customise with any ggplot2 function
or save with \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. Printing a kggplot does this
implicitly.
}
\examples{
p <- kggplot(iris, "Sepal.Length", "Sepal.Width", color = "Species")
g <- as_ggplot(p)
class(g)

}
