% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kdata.R
\name{as_kdata}
\alias{as_kdata}
\alias{as_kdata.data.frame}
\alias{as_kdata.ts}
\alias{as_kdata.matrix}
\alias{as_kdata.numeric}
\alias{as_kdata.logical}
\alias{as_kdata.character}
\alias{as_kdata.factor}
\alias{as_kdata.list}
\alias{as_kdata.default}
\title{Convert an input object to a data frame for kggplot}
\usage{
as_kdata(x, ...)

\method{as_kdata}{data.frame}(x, ...)

\method{as_kdata}{ts}(x, ...)

\method{as_kdata}{matrix}(x, ...)

\method{as_kdata}{numeric}(x, ...)

\method{as_kdata}{logical}(x, ...)

\method{as_kdata}{character}(x, ...)

\method{as_kdata}{factor}(x, ...)

\method{as_kdata}{list}(x, ...)

\method{as_kdata}{default}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Unused, for method extensions.}
}
\value{
A data frame.
}
\description{
\code{as_kdata()} is the S3 entry point that turns whatever you hand to
\code{\link[=kggplot]{kggplot()}} into a plain data frame. Add a method to teach kggplot a new
input class. A method may set the attributes \code{kgg_x}, \code{kgg_y} (default
variables to plot) and \code{kgg_type} (default plot type) on the result; they
are only used when the call does not specify \code{x}, \code{y} or \code{type}.
}
\details{
Built-in methods: \code{data.frame} (and tibbles), \code{ts}/\code{mts}, \code{matrix}, atomic
vectors (numeric, character, factor, logical), \code{list} of equal-length
vectors.
}
\examples{
as_kdata(ts(c(1, 3, 2), start = 2000))
as_kdata(matrix(1:6, 3, dimnames = list(NULL, c("a", "b"))))
as_kdata(c(2, 4, 8))

}
