% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{kgg_save}
\alias{kgg_save}
\title{Save a kggplot to a file}
\usage{
kgg_save(p, filename, ...)
}
\arguments{
\item{p}{A \code{kggplot} or \code{ggplot}.}

\item{filename}{File name to create on disk.}

\item{...}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
The file name, invisibly.
}
\description{
Thin wrapper around \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} that accepts a \code{kggplot}.
}
