% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kggplot-package.R
\docType{package}
\name{kggplot-package}
\alias{kggplot-package}
\title{kggplot: Concise, Composable 'ggplot2' Plots from a Single Call}
\description{
Draw a complete 'ggplot2' graphic (data reshaping, mappings, geometry, labels, theme, palette, legend, facets) with a single short call, and compose plots with the '+' operator through the 'kggplot' S3 class. Plot types, themes and palettes live in extensible registries.
}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
